requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asn1crypto==0.24.0        # via cryptography
bottle==0.12.9
certifi==2018.4.16        # via requests
cffi==1.11.5              # via cryptography
chardet==3.0.4            # via requests
cryptography==2.3
enum34==1.1.6             # via cryptography
idna==2.5                 # via cryptography, requests
ipaddress==1.0.16         # via cryptography
psycopg2-binary==2.7.4
pycparser==2.14           # via cffi
pycryptodome==3.6.1
pyjwt==1.7.1
pyyaml==5.1.2
requests==2.18.4
six==1.10.0
sqlalchemy==1.3.8
urllib3==1.22             # via requests
uwsgi==2.0.18
uwsgidecorators==1.1.0

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.2.0
